Revert "Made requested fixes"
This reverts commit 93d317bfb518d0ad0457f101f3baa83d3eff0ae2.

+++ zcm/transport/udpm/buffers.hpp
@@ -171,7 +171,7 @@ struct MessagePool
 
   private:
     void _freeMessageBuffer(Message *b);
-    void _removeFragBuf(size_t index);
+    void _removeFragBuf(unsigned int index);
 
   private:
     MemPool mempool;
